app/requirements.txt
APScheduler==3.7.0
certifi==2021.5.30
chardet==4.0.0
enum34==1.1.10
idna==2.10
pymongo==3.11.4
python-dotenv==0.17.1
pytz==2021.1
requests==2.25.1
six==1.16.0
tzlocal==2.1
urllib3==1.26.5
vk-api==11.9.4
